Improve edits protocol for bidirectional string properties
Editing is now the only way to change bidirectional string
properties. Edit is begun with startEdit on property, called
either on focus acquisition by the TextField or directly from
a setter, and it is finished with finishEdit, again either because
of loss of focus, or through direct call by the setter.